+++ src/index.js
@@ -1,6 +1,5 @@
 export { Authentication } from './authentication/handshake'
 export { ConnectionStatusListener } from './connection/connection-status'
-export { TransportTypes } from './connection/cometd'
 
 export { Client } from './client/basic'
 export { SmartClient } from './client/smart'
@@ -12,4 +11,4 @@ export { services } from './mapping'
  * SDK Version
  * @type {string}
  */
-export const VERSION = '3.1.3'
+export const VERSION = '3.2.0-alpha.2'